Changes: 	- PRVO and mujava.openjava.extension.EvaluationShuttle used the new methods in mujava.op.util.Mutator to check the method to mutate. 	- PRVO now checks for expected arguments (if given) when obtaining the method return type 	  of the method under consideration defined in the Api.
Fixes:
	- All operators where being used even when defining specific ones in the .properties file.
	- PNC operator was throwing an exception when dealing with a class with generics (this issue may appear in other operators and will be fixed if found).
	- A bug in mujava.util.Config#isValidOp(String) that was returning the first of the mutant operators regardless of the operator passed as argument.

Additions:

	- if the output folder exists a confirmation will be displayed before deleting that folder.
	- added new methods to mujava.op.util.Mutator to check the method to mutate :
		* checkApiConstructorNodeAgainstMethodToMutate
			checks whether a constructor matches the method to mutate defined in the Api
		* checkApiMethodNodeAgainstMethodToMutate
			checks whether a constructor matches the method to mutate defined in the Api
		* checkMethodNodeAgainstMethodToMutate
			checks whether a method mathes a expected name and arguments
		* checkConstructorNodeAgainstMethodToMutate
			checks whether a constructor mathes a expected name and arguments
	- [WIP] support for defining inner classes in .properties file.
	- [WIP] support for method names of the form NAME(TYPE*) to select exactly which method to mutate.
	- Checks in PRVO_WRITER to detect cases where a node (original or mutant) does not match the contemplated node types.
	- Debugging checks to mujava.op.util.MutantCodeWriter to detect if a mutation is not written

+++ src/mujava/app/Main.java
@@ -6,6 +6,7 @@ import java.util.Arrays;
 import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Scanner;
 import java.util.Set;
 
 import org.apache.commons.cli.CommandLine;
@@ -20,6 +21,7 @@ import org.apache.commons.cli.ParseException;
 import mujava.api.Api;
 import mujava.api.Configuration;
 import mujava.api.Mutant;
+import mujava.api.MutantsInformationHolder;
 import mujava.loader.Reloader;
 import mujava.op.PRVO;
 import mujava.util.Config;
@@ -107,7 +109,7 @@ public class Main {
 		//===========================validate mutation operators========================================//
 		List<Mutant> ops = new LinkedList<Mutant>();
 		Set<Mutant> operators = config.operators();
-		if (operators.isEmpty()) {
+		if (!operators.isEmpty()) {
 			ops.addAll(operators);
 		} else {
 			ops.addAll(mi.allOps());
@@ -157,7 +159,7 @@ public class Main {
 		System.out.println("Separator: "+Core.SEPARATOR);
 		System.out.println("Class to mutate as path: "+classToMutate.replaceAll("\\.", Core.SEPARATOR));
 		System.out.println("Class to mutate full path: "+originalSourceDir+classToMutate.replaceAll("\\.", Core.SEPARATOR)+".java\n");
-		if (!verifyFile(originalSourceDir+classToMutate.replaceAll("\\.", Core.SEPARATOR)+".java")) {
+		if (!verifyClassFile(originalSourceDir+classToMutate.replaceAll("\\.", Core.SEPARATOR)+".java")) {
 			System.err.println("Class to mutate ("+(originalSourceDir+classToMutate.replaceAll("\\.", Core.SEPARATOR)+".java")+") doesn't exist");
 			return;
 		}
@@ -554,6 +556,9 @@ public class Main {
 	private static void deleteDirectory(String dir) {
 		File directory = new File(dir);
 		if (directory.exists() && directory.isDirectory()) {
+			if (!confirmDirectoryCleaning(directory)) {
+				return;
+			}
 			deleteDirRecursively(directory);
 		}
 	}
@@ -568,7 +573,32 @@ public class Main {
 	    file.delete();
 	}
 	
+	private static boolean confirmDirectoryCleaning(File file) {
+		String input;
+		System.out.print("Please confirm that you want to delete everything inside " + file.getAbsolutePath() + " (Yes|Y|S|Si|N|No) : ");
+		Scanner in = new Scanner(System.in);
+		input = in.next();
+		if (input.toLowerCase().matches("(yes|y|si|s)")) {
+			return true;
+		}
+		if (input.toLowerCase().matches("(no|n)")) {
+			return false;
+		}
+		System.out.println();
+		System.out.println("Invalid option ("+ input + ")");
+		return confirmDirectoryCleaning(file);
+	}
+
 	private static boolean verifyFile(String f) {
+		File file = new File(f);
+		return file.exists() && file.isFile();
+	}
+	
+	private static boolean verifyClassFile(String f) {
+		int firstDolar = f.indexOf('$');
+		if (firstDolar > 0) {
+			f = f.substring(0, firstDolar) + ".java";
+		}
 		File file = new File(f);
 		return file.exists() && file.isFile();
 	}
